www/react: Move useTopbarActions hook to buildbot-ui package
