+++ www/react-base/src/views/BuildRequestView/BuildRequestView.tsx
@@ -30,11 +30,9 @@ import {
 } from "buildbot-data-js";
 import {useNavigate, useParams, useSearchParams} from "react-router-dom";
 import {Tab, Tabs} from "react-bootstrap";
-import {useTopbarItems} from "buildbot-ui";
+import {TopbarAction, useTopbarActions, useTopbarItems} from "buildbot-ui";
 import {RawData} from "../../components/RawData/RawData";
-import {TopbarAction} from "../../components/TopbarActions/TopbarActions";
 import {StoresContext} from "../../contexts/Stores";
-import {useTopbarActions} from "../../stores/TopbarActionsStore";
 import {AlertNotification} from "../../components/AlertNotification/AlertNotification";
 import {BuildSummary} from "../../components/BuildSummary/BuildSummary";
 import {PropertiesTable} from "../../components/PropertiesTable/PropertiesTable";
@@ -132,7 +130,7 @@ export const BuildRequestView = observer(() => {
   };
 
   const actions = buildTopbarActions(builder, buildRequest, isCancelling, cancelBuildRequest);
-  useTopbarActions(stores.topbarActions, actions);
+  useTopbarActions(stores.topbar, actions);
 
   const buildTabs = buildsQuery.array.map(build => (
     <Tab key={build.id} eventKey={`build-${build.number}`} title={`build ${build.number}`}>
